updated the solution to use double values instead of integers

+++ src/AskingQuestions.java
@@ -4,16 +4,21 @@ public class AskingQuestions
 {
     public static void main( String[] args )
     {
+        // Create a scanner object
         Scanner keyboard = new Scanner(System.in);
 
-        int num1, num2;
+        // Declare variables
+        double num1, num2;
 
+        // Prompt the user for the first temperature and store it
         System.out.print( "First temperature? " );
-        num1 = keyboard.nextInt();
+        num1 = keyboard.nextDouble();
 
+        // Prompt the user for the second temperature and store it
         System.out.print( "Second temperature? " );
-        num2 = keyboard.nextInt();
+        num2 = keyboard.nextDouble();
 
+        // Display the average temperature to the user
         System.out.println("The average value is : " + (num1 + num2)/2);
     }
 }
